+++ tencentcloud/rce/v20201103/models.go
@@ -55,7 +55,8 @@ type InputManageMarketingRisk struct {
 	// 账号信息。
 	Account *AccountInfo `json:"Account,omitempty" name:"Account"`
 
-	// 场景code
+	// 场景code;
+	// e_activity_antirush；活动防刷场景e_login_protection；登陆保护场景e_register_protection：注册保护场景
 	SceneCode *string `json:"SceneCode,omitempty" name:"SceneCode"`
 
 	// 登录来源的外网IP
@@ -111,6 +112,67 @@ type InputManageMarketingRisk struct {
 	DeviceType *int64 `json:"DeviceType,omitempty" name:"DeviceType"`
 
 	// 详细信息
+	// FieldName 清单
+	// Android serial_number String 否 设备序列号
+	// Android carrier String 否 运营商；-1: 获取失败，0: 其他，1: 移动，2: 联通，3: 电信，4: 铁通
+	// Android mcc_mnc String 否 netOperator MCC+MNC
+	// Android model String 否 手机型号
+	// Android os_system String 否 操作系统
+	// Android vendor_id String 否 设备品牌 “华为”“oppo”“小米”
+	// Android device_version String 否 设备版本
+	// Android android_api_level String 否 安卓API等级
+	// Android phone_chip_info String 否 手机芯片信息
+	// Android resolution_w String 否 屏幕分辨率宽，保留整数
+	// Android resolution_h String 否 屏幕分辨率高，保留整数
+	// Android brightness String 否 屏幕亮度
+	// Android bluetooth_address String 否 蓝牙地址
+	// Android baseband_version String 否 基带版本
+	// Android kernel_version String 否 kernel 版本
+	// Android cpu_core String 否 CPU 核数
+	// Android cpu_model String 否 CPU 型号
+	// Android memory String 否 内存容量，单位转换为 GB
+	// Android storage String 否 存储容量，单位转换为 GB
+	// Android volume String 否 手机音量
+	// Android battery_power String 否 电池电量
+	// Android language String 否 语言
+	// Android package_name String 否 软件包名
+	// Android app_version String 否 app 版本号
+	// Android app_name String 否 app 显示名称
+	// Android is_debug String 否 是否 debug；0 为正常模式，1 为 debug 模式；其他值无效
+	// Android is_root String 否 是否越狱；0 为正常，1 为越狱；其他值无效
+	// Android is_proxy String 否 是否启动代理；0 为未开启，1 为开启；其他值无效
+	// Android is_emulator String 否 是否模拟器；0 为未开启，1 为开启；其他值无效
+	// Android charge_status String 否 充电状态；1-不在充电，2-USB充电，3-电源充电
+	// Android network_type String 否 网络类型：2G/3G/4G/5G/WiFi/WWAN/other
+	// Android wifi_mac String 否 WiFi MAC地址
+	// IOS model String 否 机器型号 iPhone11
+	// IOS memory String 否 内存容量，单位转换为 GB
+	// IOS os_system String 否 操作系统
+	// IOS device_version String 否 设备版本
+	// IOS phone_chip_info String 否 手机芯片信息
+	// IOS device_name String 否 设备名称 "xxx 的 iPhone"，"xxx's IPhone" 等等
+	// IOS uptime String 否 开机时间
+	// IOS language String 否 系统语言
+	// IOS carrier String 否 运营商
+	// IOS cpu_model String 否 CPU 型号
+	// IOS cpu_core String 否 CPU 个数
+	// IOS volume String 否 手机音量
+	// IOS battery_power String 否 电池电量
+	// IOS resolution_w String 否 屏幕分辨率宽，保留整数
+	// IOS resolution_h String 否 屏幕分辨率高，保留整数
+	// IOS package_name String 否 app 包名
+	// IOS app_version String 否 app 版本号
+	// IOS app_name String 否 app 显示名称
+	// IOS is_debug String 否 是否 debug；0 为正常模式，1 为 debug 模式；其他值无效
+	// IOS is_root String 否 是否越狱；0 为正常，1 为越狱；其他值无效
+	// IOS is_proxy String 否 是否启动代理；0 为未开启，1 为开启；其他值无效
+	// IOS is_emulator String 否 是否模拟器；0 为未开启，1 为开启；其他值无效
+	// IOS charge_status String 否 充电状态；1-不在充电，2-USB充电，3-电源充电
+	// IOS network_type String 否 网络类型：2G/3G/4G/5G/WiFi/WWAN/other
+	// IOS wifi_mac String 否 WiFi MAC地址
+	// 其他 os_system String 否 操作系统
+	// 其他 browser String 否 浏览器信息
+	// 其他 from_url String 否 来源链接
 	Details []*InputDetails `json:"Details,omitempty" name:"Details" list`
 
 	// 可选填写。详情请跳转至SponsorInfo查看。
@@ -249,6 +311,31 @@ type OutputManageMarketingRiskValue struct {
 	RiskLevel *string `json:"RiskLevel,omitempty" name:"RiskLevel"`
 
 	// 风险类型，请参考官网风险类型
+	// 账号风险 
+	// 1 账号信用低,账号近期存在因恶意被处罚历史，网络低活跃，被举报等因素
+	// 11 疑似 低活跃账号,账号活跃度与正常用户有差异
+	// 2 垃圾账号 疑似批量注册小号，近期存在严重违规或大量举报
+	// 21 疑似小号 账号有疑似线上养号，小号等行为
+	// 22 疑似违规账号 账号曾有违规行为、曾被举报过、曾因违规被处罚过等
+	// 3 无效账号 送检账号参数无法成功解析，请检查微信 openid 是否有误/QQopenid是否与QQ开发平台的appid关联，手机号是否为中国大陆手机号 。
+	// 4 黑名单 该账号在业务侧有过拉黑记录
+	// 5 白名单 业务自行有添加过白名单记录
+	// 行为风险 
+	// 101 批量操作 存在 ip/设备/环境等因素的聚集性异常
+	// 1011 疑似 IP 属性聚集，出现 IP 聚集
+	// 1012 疑似 设备属性聚集 出现设备聚集
+	// 102 自动机 疑似自动机批量请求
+	// 103 恶意行为-网赚 疑似网赚
+	// 104 微信登录态无效 检查 WeChatAccessToken 参数，是否已经失效；
+	// 201 环境风险 环境异常 操作 ip/设备/环境存在异常。当前 ip 为非常用 ip 或恶意 ip 段
+	// 2011 疑似 非常用IP 请求 当前请求 IP 非该账号常用 IP
+	// 2012 疑似 IP 异常 使用 idc 机房 ip 或 使用代理 ip 或 使用恶意 ip 等
+	// 205 非公网有效ip 传进来的 IP 地址为内网 ip 地址或者 ip 保留地址；
+	// 设备风险
+	// 206  设备异常 该设备存在异常的使用行为
+	// 2061 疑似 非常用设备 当前请求的设备非该账号常用设备
+	// 2062 疑似 虚拟设备 请求设备为模拟器、脚本、云设备等虚拟设备
+	// 2063 疑似 群控设备 请求设备为猫池、手机墙等群控设备
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	RiskType []*int64 `json:"RiskType,omitempty" name:"RiskType" list`
 }