+++ tencentcloud/tke/v20180525/models.go
@@ -63,7 +63,7 @@ type AddExistedInstancesRequest struct {
 	// 实例列表，不支持竞价实例
 	InstanceIds []*string `json:"InstanceIds,omitempty" name:"InstanceIds" list`
 
-	// 实例额外需要设置参数信息
+	// 实例额外需要设置参数信息(默认值)
 	InstanceAdvancedSettings *InstanceAdvancedSettings `json:"InstanceAdvancedSettings,omitempty" name:"InstanceAdvancedSettings"`
 
 	// 增强服务。通过该参数可以指定是否开启云安全、云监控等服务。若不指定该参数，则默认开启云监控、云安全服务。